Update frontend to typescript & bootstrap
