+++ client/package.json
@@ -3,22 +3,41 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@types/jest": "^26.0.14",
+    "@types/node": "^14.11.2",
+    "@types/react": "^16.9.50",
+    "@types/react-dom": "^16.9.8",
     "axios": "^0.18.0",
     "bootstrap": "^4.5.2",
-    "d3": "^5.12.0",
-    "google-map-react": "^1.1.4",
-    "react": "^16.5.0",
+    "d3": "^5.16.0",
+    "google-map-react": "^1.1.7",
+    "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
-    "react-d3-speedometer": "^0.7.1",
-    "react-dom": "^16.5.0",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "1.1.5"
+    "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "3.4.3",
+    "typescript": "^4.0.3"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   },
   "proxy": "http://localhost:8080"
 }
